% --- Source file: gene_SNP_data.Rd ---
\name{gene_SNP_data}
\alias{gene_SNP_data}
%\docType{package}
\title{
 Gene-SNP data 
}
\description{
 Gene-SNP data file for \code{\link{ARTP_pathway}}
}
\details{ gene_SNP_data.txt is a tab delimited file that contains the gene that each SNP belongs to.  
 
}
\examples{
 # Load and print the first 5 rows 
 data(gene_SNP_data, package="ARTP") 

 gene_SNP_data[1:5, ]

}
\keyword{data}

